# docs/swagger.yaml
definitions:
...
  model.DIDFromPubKeyResponse:
    properties:
      did:
        type: string
    type: object
...
  server.DIDFromPubKeySwaggoRequest:
    properties:
      public_key:
        type: string
    type: object
...
info:
  contact:
    email: murali.c@ensurity.com
    name: API Support
  description: Rubix core API to control & manage the node.
  title: Rubix Core
  version: "0.9"
paths:
...
  /api/request-did-for-pubkey:
    post:
      consumes:
      - application/json
      description: This API will returns DID for corresponding public key
      operationId: request-did-for-pubkey
      parameters:
      - description: Get DID from Public Key
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/server.DIDFromPubKeySwaggoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DIDFromPubKeyResponse'
      summary: Returns DID for corresponding public key
      tags:
      - Account
...
securityDefinitions:
  SessionToken:
    in: header
    name: Session-Token
    type: apiKey
swagger: "2.0"